requirements.txt
fastapi
python-dotenv
groq
sentence-transformers
joblib
pandas
scikit-learn
xxhash
aiohttp
tf-keras
python-multipart